Startet implementing domain services to resolve dependency cycle
