+++ WaaS/WaaS.Business/Services/ScrapeJobEventService.cs
@@ -10,29 +10,30 @@ using WaaS.Business.Dtos;
 using WaaS.Business.Entities;
 using WaaS.Business.Interfaces.Repositories;
 using WaaS.Business.Interfaces.Services;
+using WaaS.Business.Interfaces.Services.Domain;
+using WaaS.Business.Services.Domain;
 
 namespace WaaS.Business.Services
 {
   public class ScrapeJobEventService : IScrapeJobEventService
   {
-
-    private readonly IScrapeJobEventRepository _scrapeJobEventRepository;
     private readonly IMapper _mapper;
     private readonly UserManager<IdentityUser> _userManager;
     private readonly IScrapeJobService _scrapeJobService;
+    private readonly IScrapeJobEventDomainService _scrapeJobEventDomainService;
 
     public ScrapeJobEventService
       (
-      IMapper mapper,
-      IScrapeJobEventRepository scrapeJobEventRepository,
-      UserManager<IdentityUser> userManager,
-      IScrapeJobService scrapeJobService
+        IMapper mapper,
+        UserManager<IdentityUser> userManager,
+        IScrapeJobService scrapeJobService,
+        IScrapeJobEventDomainService scrapeJobEventDomainService
       )
     {
       _mapper = mapper;
-      _scrapeJobEventRepository = scrapeJobEventRepository;
       _userManager = userManager;
       _scrapeJobService = scrapeJobService;
+      _scrapeJobEventDomainService = scrapeJobEventDomainService;
     }
 
 
@@ -45,7 +46,7 @@ namespace WaaS.Business.Services
 
         var entity = _mapper.Map<ScrapeJobEvent>(scrapeJobEvent);
 
-        var success = await Create(entity);
+        var success = await _scrapeJobEventDomainService.Create(entity);
 
         if (success)
         {
@@ -57,24 +58,14 @@ namespace WaaS.Business.Services
       return null;
     }
 
-    public async Task<bool> Create(ScrapeJobEvent scrapeJobEvent)
-    {
-      var result = false;
-      if (scrapeJobEvent != null)
-      {
-        result = await _scrapeJobEventRepository.AddAsync(scrapeJobEvent);
-      }
-      return result;
-    }
-
     public async Task<bool> Delete(long id, ClaimsPrincipal principal)
     {
       var idUser = await _userManager.GetUserAsync(principal);
-      var entity = await _scrapeJobEventRepository.GetAsync(id);
+      var entity = await _scrapeJobEventDomainService.GetAsync(id);
 
       if (entity != null && await _scrapeJobService.ScrapeJobIsOfUser(entity.ScrapeJob.Id, idUser.Id))
       {
-        return await _scrapeJobEventRepository.DeleteAsync(id);
+        return await _scrapeJobEventDomainService.DeleteAsync(id);
       }
 
       return false;
@@ -84,7 +75,7 @@ namespace WaaS.Business.Services
     public async Task<ScrapeJobEventDto> Read(long id, ClaimsPrincipal principal)
     {
       var idUser = await _userManager.GetUserAsync(principal);
-      var entity = await _scrapeJobEventRepository.GetAsync(id);
+      var entity = await _scrapeJobEventDomainService.GetAsync(id);
 
       if (entity != null && await _scrapeJobService.ScrapeJobIsOfUser(entity.ScrapeJob.Id, idUser.Id))
       {
@@ -101,7 +92,7 @@ namespace WaaS.Business.Services
 
       if(scrapeJob != null)
       {
-        var entities = _scrapeJobEventRepository.ReadScrapeJobEventsOfScrapeJob(scrapeJob.Id);
+        var entities = _scrapeJobEventDomainService.ReadScrapeJobEventsOfScrapeJob(scrapeJob.Id);
 
         if (entities.Any())
         {
@@ -116,15 +107,15 @@ namespace WaaS.Business.Services
     public async Task<ScrapeJobEventDto> Update(ScrapeJobEventDto scrapeJobEvent, ClaimsPrincipal principal)
     {
       var idUser = await _userManager.GetUserAsync(principal);
-      var entity = await _scrapeJobEventRepository.GetAsync(scrapeJobEvent.Id);
+      var entity = await _scrapeJobEventDomainService.GetAsync(scrapeJobEvent.Id);
 
       if (entity != null && await _scrapeJobService.ScrapeJobIsOfUser(entity.ScrapeJob.Id, idUser.Id))
       {
-        var success = await _scrapeJobEventRepository.UpdateAsync(scrapeJobEvent.Id, e => e = _mapper.Map(scrapeJobEvent, e));
+        var success = await _scrapeJobEventDomainService.UpdateAsync(scrapeJobEvent.Id, e => e = _mapper.Map(scrapeJobEvent, e));
 
         if (success)
         {
-          var updatedEntity = await _scrapeJobEventRepository.GetAsync(scrapeJobEvent.Id);
+          var updatedEntity = await _scrapeJobEventDomainService.GetAsync(scrapeJobEvent.Id);
           return _mapper.Map<ScrapeJobEventDto>(updatedEntity);
         }
       }
